Rearrange code
1. Distinct subfolders nginx_server/ server/ for
NGINX and builtin servers.
2. Extract common (not specific to particular
tsgi adapter) code to http/tsgi.lua.
